Trigger counters are of type uintmax_t, not size_t.
git-svn-id: http://svn.jdtaus.org/svnroot/jdtaus/Abagnale/trunk@9234 d3f52c91-230f-0410-802c-e3ec080d4bae

+++ config.h
@@ -72,9 +72,9 @@ struct MarketConfig {
   struct Numeric *restrict sl_dlnanos;
   struct Numeric *restrict tl_dlnanos;
   struct Numeric *restrict tp_dlnanos;
-  size_t sl_dlcnt;
-  size_t tl_dlcnt;
-  size_t tp_dlcnt;
+  uintmax_t sl_dlcnt;
+  uintmax_t tl_dlcnt;
+  uintmax_t tp_dlcnt;
 };
 
 void config_init(void);
